Use content language instead of user language
This extension translates the webhook messages according to the UI
language of the user who is making the change. This may not always be
correct - the wiki I edit and the connected Discord server are fully
in English, but since I have the wiki's interface language set to my
native language, the webhook messages were being sent with the
translations hard-baked into them to an English-speaking community.

This commit adds ->inContentLanguage() to all the wfMessage()s so the
webhook messages will be in the same language as the wiki content.
(See https://www.mediawiki.org/wiki/Manual:Messages_API#Language)